# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Packwerk B2B Packaging Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/orders:
    get:
      operationId: getDashboardOrders
      tags: [orders]
      summary: Get all orders for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /dashboard/profile:
    get:
      operationId: getDashboardProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateDashboardProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /admin/quotes:
    get:
      operationId: adminListQuotes
      tags: [admin]
      summary: List all quotes (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quotes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteRequest"
  /admin/quotes/{id}/status:
    put:
      operationId: adminUpdateQuoteStatus
      tags: [admin]
      summary: Update quote status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [submitted, reviewing, quoted, accepted, rejected]
                rejection_reason:
                  type: string
              required: [status]
      responses:
        "200":
          description: Updated
  /admin/quotes/{id}/accept:
    post:
      operationId: adminAcceptQuote
      tags: [admin]
      summary: Accept quote, create order and user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                total_price:
                  type: number
                payment_type:
                  type: string
                  enum: [upfront, standard, credit]
                line_items:
                  type: array
                  items:
                    type: object
              required: [total_price, payment_type, line_items]
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id:
                    type: string
                  user_created:
                    type: boolean
                required: [order_id, user_created]
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}/status:
    put:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [confirmed, in_production, qc, dispatched, delivered]
                tracking_number:
                  type: string
                tracking_url:
                  type: string
              required: [status]
      responses:
        "200":
          description: Updated
  /admin/designs:
    get:
      operationId: adminListDesigns
      tags: [admin]
      summary: List all design requests (admin)
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Design requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DesignRequest"
  /admin/designs/{id}/status:
    put:
      operationId: adminUpdateDesignStatus
      tags: [admin]
      summary: Update design request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [paid, in_design, revision, approved]
                designer_notes:
                  type: string
              required: [status]
      responses:
        "200":
          description: Updated
  /admin/samples:
    get:
      operationId: adminListSamples
      tags: [admin]
      summary: List all sample requests (admin)
      parameters:
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sample requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SampleRequest"
  /admin/samples/{id}/dispatch:
    post:
      operationId: adminDispatchSample
      tags: [admin]
      summary: Mark sample as dispatched
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated
  /auth/login:
    post:
      operationId: login
      tags: [users]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required: [email, password]
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserProfile"
                required: [access_token, user]
        "401":
          description: Invalid credentials
...
components:
  schemas:
...
    QuoteRequest:
      type: object
      properties:
        id:
          type: string
        quote_id:
          type: string
        contact_name:
          type: string
        company_name:
          type: string
        email:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            type: object
        delivery_country:
          type: string
        delivery_pincode:
          type: string
          nullable: true
        preferred_timeline:
          type: string
        notes:
          type: string
          nullable: true
        total_estimated_min:
          type: number
          nullable: true
        total_estimated_max:
          type: number
          nullable: true
        status:
          type: string
          enum: [submitted, reviewing, quoted, accepted, rejected]
        rejection_reason:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, quote_id, contact_name, company_name, email, phone, items, delivery_country, preferred_timeline, status]
    Order:
      type: object
      properties:
        id:
          type: string
        order_id:
          type: string
        items:
          type: array
          items:
            type: object
        total_price:
          type: number
        payment_type:
          type: string
        discount_applied:
          type: number
        delivery_address:
          type: object
        status:
          type: string
          enum: [confirmed, in_production, qc, dispatched, delivered]
        estimated_delivery:
          type: string
          nullable: true
        tracking_number:
          type: string
          nullable: true
        tracking_url:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, order_id, items, total_price, payment_type, discount_applied, delivery_address, status, created_at]
    DesignRequest:
      type: object
      properties:
        id:
          type: string
        design_id:
          type: string
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
        product_type:
          type: string
        brand_colors:
          type: string
          nullable: true
        brand_description:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        is_rush:
          type: boolean
        amount_paid:
          type: integer
        status:
          type: string
          enum: [paid, in_design, revision, approved]
        designer_notes:
          type: string
          nullable: true
        output_files:
          type: array
          items:
            type: string
          nullable: true
        created_at:
          type: string
      required: [id, design_id, contact_name, email, phone, product_type, is_rush, amount_paid, status, created_at]
...
    SampleRequest:
      type: object
      properties:
        id:
          type: string
        sample_id:
          type: string
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
        product_id:
          type: string
        sample_tier:
          type: string
          enum: [standard, premium, complex]
        amount_paid:
          type: integer
        status:
          type: string
          enum: [paid, dispatched, delivered]
        created_at:
          type: string
      required: [id, sample_id, contact_name, email, phone, product_id, sample_tier, amount_paid, status, created_at]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        company_name:
          type: string
          nullable: true
        contact_name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        gstin:
          type: string
          nullable: true
        country:
          type: string
        orders_completed:
          type: integer
        credit_eligible:
          type: boolean
        credit_limit:
          type: number
        default_address:
          type: object
          nullable: true
      required: [id, email, country, orders_completed, credit_eligible, credit_limit]
    UpdateProfileRequest:
      type: object
      properties:
        company_name:
          type: string
        contact_name:
          type: string
        phone:
          type: string
        gstin:
          type: string
        default_address:
          type: object
...