# lib/api-zod/src/generated/types/quoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Packwerk B2B Packaging Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteRequestItemsItem } from "./quoteRequestItemsItem";
import type { QuoteRequestStatus } from "./quoteRequestStatus";

export interface QuoteRequest {
  id: string;
  quote_id: string;
  contact_name: string;
  company_name: string;
  email: string;
  phone: string;
  items: QuoteRequestItemsItem[];
  delivery_country: string;
  delivery_pincode?: string | null;
  preferred_timeline: string;
  notes?: string | null;
  total_estimated_min?: number | null;
  total_estimated_max?: number | null;
  status: QuoteRequestStatus;
  rejection_reason?: string | null;
  created_at?: string;
}