# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Packwerk B2B Packaging Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProfileRequestDefaultAddress } from "./updateProfileRequestDefaultAddress";

export interface UpdateProfileRequest {
  company_name?: string;
  contact_name?: string;
  phone?: string;
  gstin?: string;
  default_address?: UpdateProfileRequestDefaultAddress;
}